prevent deletion of readme.md

+++ updater/lib/updater_functions.php
@@ -216,7 +216,10 @@ $files = glob($foldername . '/*');
 
 	foreach ($files as $file) {
 		if (is_file($file)) {
-			unlink($file);
+			if ($file != "readme.md") {
+				unlink($file);
+			}
+			
 		}
 	}
 }
